+++ packages/cookie/src/lib/cookie.ts
@@ -6,6 +6,16 @@ import {
 
 import { sign, unsign } from './crypto.ts'
 
+/**
+ * Creates a new `Cookie` object.
+ * @param name The name of the cookie
+ * @param options The options for the cookie
+ * @returns A new cookie instance
+ */
+export function createCookie(name: string, options?: CookieOptions): Cookie {
+  return new Cookie(name, options)
+}
+
 export interface CookieOptions {
   /**
    * A function that decodes the cookie value.
@@ -41,11 +51,6 @@ export interface CookieOptions {
  * to sign/unsign the value of the cookie to ensure it's not tampered with.
  */
 export class Cookie {
-  readonly name: string
-  readonly #decode: (value: string) => string
-  readonly #encode: (value: string) => string
-  readonly #secrets: string[]
-
   constructor(name: string, options?: CookieOptions) {
     this.name = name
     this.#decode = options?.decode ?? decodeURIComponent
@@ -54,9 +59,18 @@ export class Cookie {
   }
 
   /**
+   * The name of the cookie.
+   */
+  readonly name: string
+
+  readonly #decode: (value: string) => string
+  readonly #encode: (value: string) => string
+  readonly #secrets: string[]
+
+  /**
    * True if this cookie uses one or more secrets for verification.
    */
-  get isSigned(): boolean {
+  get signed(): boolean {
     return this.#secrets.length > 0
   }